requirements.txt
st7789
pillow
rpi-lgpio
gpiozero
pygame
pyfluidsynth